refactor: split actions into separate modules
