+++ src/__init__.py
@@ -11,6 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from . import app
+from . import actions, app
 
-__ALL__ = [app]
+__ALL__ = [app, actions]
