+++ src/app.py
@@ -11,16 +11,11 @@
 # either express or implied. See the License for the specific language governing permissions
 # and limitations under the License.
 
-import hashlib
-import json
 import os
-import re
 import ssl
 import sys
 import time
-from collections.abc import Iterator
-from datetime import datetime, UTC
-from dateutil.parser import ParserError, parse as dateutil_parse
+from datetime import UTC, datetime
 from io import BytesIO
 from urllib.error import HTTPError as UrllibHTTPError, URLError
 from urllib.request import ProxyHandler, Request, build_opener, install_opener, urlopen
@@ -33,33 +28,20 @@ import splunklib.results as splunk_results
 from splunklib.binding import HTTPError as SplunkHTTPError
 import xmltodict
 from bs4 import BeautifulSoup
-from bs4.dammit import UnicodeDammit
 from soar_sdk.abstract import SOARClient
-from soar_sdk.action_results import ActionOutput, OutputField, PermissiveActionOutput
 from soar_sdk.app import App
 from soar_sdk.asset import AssetField, BaseAsset, FieldCategory
 from soar_sdk.logging import getLogger
-from soar_sdk.models.artifact import Artifact
-from soar_sdk.models.container import Container
-from soar_sdk.params import OnPollParams, Param, Params
 
 from .splunk_consts import (
-    CIM_CEF_MAP,
     SPLUNK_DEFAULT_REQUEST_TIMEOUT,
-    SPLUNK_DEFAULT_SOURCE,
-    SPLUNK_DEFAULT_SOURCE_TYPE,
-    SPLUNK_DISPOSITION_QUERY_FORMAT,
-    SPLUNK_ERR_BAD_DISPOSITION,
-    SPLUNK_ERR_BAD_STATUS,
     SPLUNK_ERR_CONNECTIVITY_FAILED,
     SPLUNK_ERR_CONNECTIVITY_TEST,
     SPLUNK_ERR_EMPTY_RESPONSE,
     SPLUNK_ERR_INVALID_INTEGER,
     SPLUNK_ERR_INVALID_SLEEP_TIME,
-    SPLUNK_ERR_NEED_PARAM,
-    SPLUNK_ERR_NON_NEGATIVE_INTEGER,
-    SPLUNK_ERR_NOT_ES,
     SPLUNK_ERR_INVALID_PARAM,
+    SPLUNK_ERR_NON_NEGATIVE_INTEGER,
     SPLUNK_ERR_REQUIRED_CONFIG_PARAMS,
     SPLUNK_ERR_SPLUNK_JOB_HAS_TIMED_OUT,
     SPLUNK_ERR_UNABLE_TO_CREATE_JOB,
@@ -70,8 +52,6 @@ from .splunk_consts import (
     SPLUNK_PROG_CREATED_QUERY,
     SPLUNK_PROG_CREATING_SEARCH_JOB,
     SPLUNK_RID_SID_NOTABLE_QUERY,
-    SPLUNK_SEARCH_MODE_SMART,
-    SPLUNK_SEVERITY_MAP,
     SPLUNK_SUCCESS_CONNECTIVITY_TEST,
 )
 
@@ -654,246 +634,6 @@ class SplunkHelper:
 
 
 # ---------------------------------------------------------------------------
-# Ingestion helpers
-# ---------------------------------------------------------------------------
-def _get_event_start(start_time: str | None) -> str | None:
-    if not start_time:
-        return None
-    try:
-        datetime_obj = dateutil_parse(start_time)
-        return datetime_obj.astimezone(UTC).strftime("%Y-%m-%dT%H:%M:%S.%fZ")
-    except ParserError as e:
-        logger.error("ParserError while parsing _time: %s", e)
-        return None
-    except Exception as e:
-        logger.error("Exception while parsing _time: %s", e)
-        return None
-
-
-def _get_fips_enabled() -> bool:
-    try:
-        from phantom_common.install_info import is_fips_enabled  # noqa: PLC0415
-        return is_fips_enabled()
-    except ImportError:
-        return False
-
-
-def _get_splunk_severity(item: dict) -> str:
-    severity = item.get("severity")
-    if isinstance(severity, list):
-        for key in ["critical", "high", "medium", "low", "informational"]:
-            if key in severity:
-                return SPLUNK_SEVERITY_MAP[key]
-        return ""
-    severity = SPLUNK_SEVERITY_MAP.get(severity) if severity else None
-    if not severity:
-        urgency = item.get("urgency")
-        severity = SPLUNK_SEVERITY_MAP.get(urgency, "medium")
-    return severity
-
-
-def _get_splunk_title(item: dict, prefix: str, name_values: list[str]) -> str:
-    title = prefix
-    values_list = list(name_values)
-    if not title and not values_list:
-        values_list.append("source")
-
-    values = ""
-    for i, nv in enumerate(values_list):
-        if CIM_CEF_MAP.get(nv) and item.get(CIM_CEF_MAP.get(nv)):
-            value = item.get(CIM_CEF_MAP.get(nv))
-        elif item.get(nv):
-            value = item.get(nv)
-        else:
-            value = CIM_CEF_MAP.get(nv, nv)
-        values += f"{value}" + ("" if i == len(values_list) - 1 else ", ")
-
-    if not title:
-        t = item.get("_time")
-        title = f"Splunk Log Entry on {t}" if t else "Splunk Log Entry"
-    else:
-        title = item.get(title, title)
-
-    return f"{title}: {values}"
-
-
-# ---------------------------------------------------------------------------
-# Params / Outputs
-# ---------------------------------------------------------------------------
-class RunQueryParams(Params):
-    command: str = Param(
-        description="Beginning command (in Splunk Processing Language)",
-        required=False,
-        value_list=["search", "eval", "savedsearch", "stats", "table", "tstats"],
-        default="search",
-    )
-    query: str = Param(
-        description="Query to run (in Splunk Processing Language)",
-        required=True,
-        primary=True,
-        cef_types=["splunk query"],
-    )
-    display: str = Param(description="Display fields (comma-separated)", required=False, default="")
-    parse_only: bool = Param(description="Parse only", required=False, default=False)
-    add_raw_field: bool = Param(description="Ingest _raw field data", required=False, default=True)
-    attach_result: bool = Param(description="Attach result to the vault", required=False, default=False)
-    start_time: str = Param(description="Earliest time modifier", required=False, default="")
-    end_time: str = Param(description="Latest time modifier", required=False, default="")
-    search_mode: str = Param(
-        description="Search mode",
-        required=False,
-        value_list=["fast", "verbose", "smart"],
-        default="smart",
-    )
-    time_format: str = Param(description="Custom timestamp format", required=False, default="")
-
-
-class RunQueryOutput(PermissiveActionOutput):
-    pass
-
-
-class RunQuerySummary(ActionOutput):
-    sid: str | None = None
-    total_events: int | None = None
-
-
-class GetHostEventsParams(Params):
-    ip_hostname: str = Param(
-        description="Hostname/IP to search the events of",
-        required=True,
-        primary=True,
-        cef_types=["ip", "host name"],
-    )
-    last_n_days: str = Param(description="Number of days ago", required=False, default="")
-
-
-class GetHostEventsOutput(PermissiveActionOutput):
-    host: str | None = OutputField(column_name="Host")
-    time: str | None = OutputField(column_name="Time", alias="_time")
-    raw: str | None = OutputField(column_name="Raw", alias="_raw")
-
-
-class GetHostEventsSummary(ActionOutput):
-    sid: str | None = None
-    total_events: int | None = None
-
-
-class UpdateEventParams(Params):
-    event_ids: str = Param(
-        description="Event ID to update",
-        required=True,
-        primary=True,
-        cef_types=["splunk notable event id"],
-    )
-    owner: str = Param(description="New owner for the event", required=False, default="")
-    status: str = Param(
-        description="New status for the event",
-        required=False,
-        default="",
-        value_list=["", "unassigned", "new", "in progress", "pending", "resolved", "closed"],
-    )
-    integer_status: str = Param(description="Integer representing custom status value", required=False, default="")
-    urgency: str = Param(
-        description="New urgency for the event",
-        required=False,
-        default="",
-        value_list=["", "informational", "low", "medium", "high", "critical"],
-    )
-    comment: str = Param(description="New comment for the event", required=False, default="")
-    disposition: str = Param(
-        description="New disposition field",
-        required=False,
-        default="",
-        value_list=[
-            "", "Unassigned", "True Positive - Suspicious Activity",
-            "Benign Positive - Suspicious But Expected",
-            "False Positive - Incorrect Analytic Logic",
-            "False Positive - Inaccurate Data", "Undetermined", "Other",
-        ],
-    )
-    integer_disposition: str = Param(description="Integer representing custom disposition value", required=False, default="")
-    wait_for_confirmation: bool = Param(description="Validate event_ids", required=False, default=False)
-
-
-class UpdateEventOutput(ActionOutput):
-    status: str | None = OutputField(column_name="Status")
-    failure_count: int | None = None
-    message: str | None = OutputField(column_name="Message")
-    success: bool | None = None
-    success_count: int | None = None
-
-
-class UpdateEventSummary(ActionOutput):
-    sid: str | None = None
-    updated_event_id: str | None = None
-
-
-class PostDataParams(Params):
-    data: str = Param(description="Data to post", required=True)
-    host: str = Param(
-        description="Host for event",
-        required=False,
-        default="",
-        primary=True,
-        cef_types=["ip", "host name"],
-    )
-    index: str = Param(description="Index to send event to", required=False, default="")
-    source: str = Param(description="Source for event", required=False, default="Phantom")
-    source_type: str = Param(
-        description="Type of source for event",
-        required=False,
-        default="Automation/Orchestration Platform",
-    )
-
-
-class PostDataOutput(ActionOutput):
-    status: str | None = OutputField(column_name="Status")
-    message: str | None = OutputField(column_name="Message")
-
-
-# ---------------------------------------------------------------------------
-# Custom view for run query
-# ---------------------------------------------------------------------------
-@app.view_handler(template="splunk_run_query.html")
-def display_view(outputs: list[RunQueryOutput]) -> dict:
-    if not outputs:
-        return {"results": [{"data": {}, "param": {}}]}
-
-    first = outputs[0].model_dump(exclude_none=True)
-    param = {
-        "query": first.get("_param_query", ""),
-        "command": first.get("_param_command", ""),
-        "display": first.get("_param_display", ""),
-        "parse_only": first.get("_param_parse_only", False),
-        "search_mode": first.get("_param_search_mode", "smart"),
-    }
-    display_fields = param.get("display", "")
-
-    all_data = []
-    for output in outputs:
-        data = {k: v for k, v in output.model_dump(exclude_none=True).items() if not k.startswith("_")}
-        all_data.append(data)
-
-    if display_fields:
-        headers = [x.strip() for x in display_fields.split(",") if x.strip()]
-    elif all_data:
-        headers = [k for k in all_data[0] if not k.startswith("_")]
-    else:
-        headers = []
-
-    processed_data = [{h: item.get(h) for h in headers} for item in all_data]
-
-    return {
-        "results": [{
-            "param": param,
-            "data": all_data or {},
-            "processed_data": processed_data,
-            "headers": headers,
-        }],
-    }
-
-
-# ---------------------------------------------------------------------------
 # Test Connectivity
 # ---------------------------------------------------------------------------
 @app.test_connectivity()
@@ -916,398 +656,6 @@ def test_connectivity(soar: SOARClient, asset: Asset) -> None:
     logger.progress("Detected Splunk %sserver version %s", "ES " if is_es else "", version)
     soar.set_message(SPLUNK_SUCCESS_CONNECTIVITY_TEST)
     logger.info(SPLUNK_SUCCESS_CONNECTIVITY_TEST)
-
-
-# ---------------------------------------------------------------------------
-# Run Query
-# ---------------------------------------------------------------------------
-@app.action(
-    description="Run a search query on the Splunk device. Please escape any quotes that are part of the query string",
-    action_type="investigate",
-    read_only=True,
-    view_handler=display_view,
-    summary_type=RunQuerySummary,
-)
-def run_query(params: RunQueryParams, soar: SOARClient, asset: Asset) -> list[RunQueryOutput]:
-    helper = SplunkHelper(asset)
-    helper.validate_asset()
-    helper.connect()
-
-    search_mode = params.search_mode or SPLUNK_SEARCH_MODE_SMART
-    kwargs: dict = {"adhoc_search_level": search_mode}
-    if params.start_time:
-        kwargs["earliest_time"] = params.start_time
-    if params.end_time:
-        kwargs["latest_time"] = params.end_time
-    if params.time_format:
-        kwargs["time_format"] = params.time_format
-
-    search_command = params.command
-    search_string = params.query
-
-    if not search_command:
-        if search_string[0] != "|" and not search_string.startswith("search"):
-            search_string = f"search {search_string.strip()}"
-        search_query = search_string
-    else:
-        search_query = f"{search_command.strip()} {search_string.strip()}"
-
-    sid, results_list = helper.run_query(
-        search_query,
-        kwargs_create=kwargs,
-        parse_only=params.parse_only,
-        add_raw_field=params.add_raw_field,
-    )
-
-    if params.attach_result:
-        _attach_json_result(soar, results_list)
-
-    soar.set_summary(RunQuerySummary(sid=sid, total_events=len(results_list)))
-    soar.set_message(f"Sid: {sid}, Total events: {len(results_list)}")
-
-    param_info = {
-        "_param_query": params.query,
-        "_param_command": params.command,
-        "_param_display": params.display,
-        "_param_parse_only": params.parse_only,
-        "_param_search_mode": search_mode,
-    }
-    return [RunQueryOutput(**{**r, **param_info}) for r in results_list]
-
-
-def _attach_json_result(soar: SOARClient, data: list[dict]):
-    try:
-        container_id = soar.get_executing_container_id()
-        soar.vault.create_attachment(
-            container_id=container_id,
-            file_content=json.dumps(data),
-            file_name="splunk_run_query_result.json",
-        )
-    except Exception as e:
-        logger.error("Error attaching results to vault: %s", e)
-
-
-# ---------------------------------------------------------------------------
-# Get Host Events
-# ---------------------------------------------------------------------------
-@app.action(
-    description="Get events pertaining to a host that have occurred in the last 'N' days",
-    action_type="investigate",
-    read_only=True,
-    render_as="table",
-    summary_type=GetHostEventsSummary,
-)
-def get_host_events(params: GetHostEventsParams, soar: SOARClient, asset: Asset) -> list[GetHostEventsOutput]:
-    helper = SplunkHelper(asset)
-    helper.validate_asset()
-    helper.connect()
-
-    ip_hostname = params.ip_hostname
-    last_n_days = SplunkHelper.validate_integer(params.last_n_days, "'last_n_days' action")
-
-    search_query = f'search host="{ip_hostname}"'
-    if last_n_days:
-        search_query += f" earliest=-{last_n_days}d"
-
-    sid, results_list = helper.run_query(search_query)
-
-    soar.set_summary(GetHostEventsSummary(sid=sid, total_events=len(results_list)))
-    soar.set_message(f"Sid: {sid}, Total events: {len(results_list)}")
-
-    return [GetHostEventsOutput(**r) for r in results_list]
-
-
-# ---------------------------------------------------------------------------
-# Update Event
-# ---------------------------------------------------------------------------
-@app.action(
-    description="Update a notable event",
-    action_type="generic",
-    read_only=False,
-    render_as="table",
-    summary_type=UpdateEventSummary,
-)
-def update_event(params: UpdateEventParams, soar: SOARClient, asset: Asset) -> list[UpdateEventOutput]:
-    helper = SplunkHelper(asset)
-    helper.validate_asset()
-
-    if not helper.check_for_es():
-        raise RuntimeError(SPLUNK_ERR_NOT_ES)
-
-    ids = params.event_ids
-    owner = params.owner
-    status = params.status
-    comment = params.comment
-    urgency = params.urgency
-    disposition = params.disposition or ""
-    wait_for_confirmation = params.wait_for_confirmation
-
-    integer_status = SplunkHelper.validate_integer(
-        params.integer_status, "'integer_status' action", allow_zero=True
-    )
-    integer_disposition = SplunkHelper.validate_integer(
-        params.integer_disposition, "'integer_disposition' action", allow_zero=True
-    )
-
-    if not any([comment, status, urgency, owner, disposition]) and integer_status is None and integer_disposition is None:
-        raise ValueError(SPLUNK_ERR_NEED_PARAM)
-
-    splunk_status_dict: dict[str, int] = {}
-    splunk_disposition_dict: dict[str, int] = {}
-
-    if status or integer_status is not None:
-        splunk_status_dict = helper.get_status_dict("notable")
-        if not splunk_status_dict:
-            raise RuntimeError("Error occurred while fetching Splunk event status")
-
-    if disposition or integer_disposition is not None:
-        splunk_disposition_dict = helper.get_status_dict("disposition")
-        if not splunk_disposition_dict:
-            raise RuntimeError("Error occurred while fetching Splunk event disposition")
-
-    helper.connect()
-
-    # Resolve SID+RID combo to event_id
-    regexp = re.compile(r"\+\d*(\.\d+)?[\"$]")
-    if regexp.search(json.dumps(ids)):
-        logger.progress("Interpreting the event ID as an SID + RID combo")
-        try:
-            ids = helper.resolve_event_id(ids)
-        except Exception:
-            raise RuntimeError("Unable to find underlying event_id from SID + RID combo") from None
-
-    if wait_for_confirmation:
-        search_query = f"search `notable_by_id({ids})`"
-        _sid, validate_results = helper.run_query(search_query)
-        if not validate_results:
-            raise ValueError("Please provide a valid event ID")
-
-    request_body: dict = {"ruleUIDs": ids}
-
-    # Status
-    if integer_status is not None:
-        if int(integer_status) not in list(splunk_status_dict.values()):
-            raise ValueError(
-                "Please provide a valid value in 'integer_status' action parameter. "
-                f"Valid values: {', '.join(map(str, splunk_status_dict.values()))}"
-            )
-        request_body["status"] = str(integer_status)
-    elif status:
-        if status not in splunk_status_dict:
-            if not status.isdigit():
-                raise ValueError(SPLUNK_ERR_BAD_STATUS)
-            request_body["status"] = status
-        else:
-            request_body["status"] = splunk_status_dict[status]
-
-    # Disposition
-    if integer_disposition is not None:
-        if int(integer_disposition) not in splunk_disposition_dict.values():
-            raise ValueError(
-                "Please provide a valid value in 'integer_disposition' action parameter. "
-                f"Valid values: {', '.join(map(str, splunk_disposition_dict.values()))}"
-            )
-        request_body["disposition"] = SPLUNK_DISPOSITION_QUERY_FORMAT.format(integer_disposition)
-    elif disposition:
-        if disposition not in splunk_disposition_dict:
-            if not disposition.isdigit():
-                raise ValueError(SPLUNK_ERR_BAD_DISPOSITION)
-            request_body["disposition"] = SPLUNK_DISPOSITION_QUERY_FORMAT.format(disposition)
-        else:
-            request_body["disposition"] = SPLUNK_DISPOSITION_QUERY_FORMAT.format(splunk_disposition_dict[disposition])
-
-    param_mapping = {"urgency": urgency, "comment": comment, "newOwner": owner}
-    request_body.update({k: v for k, v in param_mapping.items() if v})
-
-    resp_data = helper.make_rest_call_retry("notable_update", request_body)
-
-    if resp_data and "success" in resp_data and not resp_data.get("success"):
-        msg = resp_data.get("message")
-        raise RuntimeError(msg if msg else "Unable to update the notable event")
-
-    soar.set_summary(UpdateEventSummary(updated_event_id=ids))
-
-    if wait_for_confirmation:
-        msg = f"Updated Event ID: {ids}"
-    else:
-        msg = (
-            f"Updated Event ID: {ids}. The event_id has not been verified. "
-            "Please confirm that the provided event_id corresponds to an actual notable event"
-        )
-
-    soar.set_message(msg)
-
-    if resp_data:
-        resp_data["status"] = "success"
-        resp_data["message"] = msg
-        return [UpdateEventOutput(**resp_data)]
-    return [UpdateEventOutput(status="success", message=msg)]
-
-
-# ---------------------------------------------------------------------------
-# Post Data
-# ---------------------------------------------------------------------------
-@app.action(
-    description="Post data to Splunk",
-    action_type="generic",
-    read_only=False,
-    render_as="table",
-)
-def post_data(params: PostDataParams, soar: SOARClient, asset: Asset) -> list[PostDataOutput]:
-    helper = SplunkHelper(asset)
-    helper.validate_asset()
-
-    try:
-        post_bytes = UnicodeDammit(params.data).unicode_markup.encode("utf-8")
-    except Exception as e:
-        logger.error("Error while encoding data: %s", e)
-        post_bytes = params.data.encode("utf-8")
-
-    get_params: dict[str, str] = {
-        "source": params.source or SPLUNK_DEFAULT_SOURCE,
-        "sourcetype": params.source_type or SPLUNK_DEFAULT_SOURCE_TYPE,
-    }
-    if params.host:
-        get_params["host"] = params.host
-    if params.index:
-        get_params["index"] = params.index
-
-    helper.make_rest_call_retry("receivers/simple", post_bytes, params=get_params)
-
-    soar.set_message("Successfully posted the data")
-    return [PostDataOutput(status="success", message="Successfully posted the data")]
-
-
-# ---------------------------------------------------------------------------
-# On Poll
-# ---------------------------------------------------------------------------
-@app.on_poll()
-def on_poll(params: OnPollParams, soar: SOARClient, asset: Asset) -> Iterator[Container | Artifact]:
-    helper = SplunkHelper(asset)
-    helper.validate_asset()
-    helper.connect()
-
-    search_command = asset.on_poll_command
-    search_string = asset.on_poll_query
-    po = asset.on_poll_parse_only
-    include_cim_fields = asset.include_cim_fields
-    use_event_id_sdi = asset.use_event_id_sdi
-
-    if not search_string:
-        raise ValueError("Need to specify Query String to use polling")
-
-    try:
-        if not search_command:
-            if search_string[0] != "|" and not search_string.startswith("search"):
-                search_string = f"search {search_string.strip()}"
-            search_query = search_string
-        else:
-            search_query = f"{search_command.strip()} {search_string.strip()}"
-    except Exception:
-        raise ValueError("Error occurred while parsing the search query") from None
-
-    search_params: dict = {}
-    state = asset.ingest_state
-    is_poll_now = params.is_manual_poll()
-
-    if is_poll_now:
-        search_params["max_count"] = params.container_count or 100
-    else:
-        search_params["max_count"] = asset.max_container
-        start_time = state.get("start_time")
-        if start_time:
-            search_params["index_earliest"] = start_time
-
-    if int(search_params["max_count"]) <= 0:
-        logger.debug("container_count <= 0, ignoring max_count")
-        search_params.pop("max_count")
-
-    try:
-        _sid, results_list = helper.run_query(search_query, kwargs_create=search_params, parse_only=po)
-    except Exception as e:
-        msg = str(e)
-        if "Invalid index_earliest" in msg:
-            logger.debug("Invalid start_time %s, retrying without it", search_params.get("index_earliest"))
-            state.pop("start_time", None)
-        raise
-
-    display = asset.on_poll_display
-    header_set = None
-    if display:
-        header_set = [x.strip().lower() for x in display.split(",")]
-
-    data = list(reversed(results_list))
-    logger.info("Total %d event(s) fetched", len(data))
-
-    container_name_prefix = asset.container_name_prefix or ""
-    raw_values = asset.container_name_values
-    container_name_values = [x.strip() for x in raw_values.split(",")] if raw_values else []
-
-    count = 1
-    for item in data:
-        try:
-            cef: dict = {}
-            if "_serial" in item:
-                item.pop("_serial")
-
-            if header_set:
-                name_mappings = {k.lower(): k for k in item if k.lower() in header_set}
-                for h in header_set:
-                    cef_name = CIM_CEF_MAP.get(h, h)
-                    cef_name = name_mappings.get(cef_name, cef_name)
-                    cef_key_value = name_mappings.get(h, h)
-                    cef[cef_name] = item.get(cef_key_value)
-                    if include_cim_fields:
-                        cef[cef_key_value] = item.get(cef_key_value)
-            else:
-                for k, v in item.items():
-                    cef[CIM_CEF_MAP.get(k, k)] = v
-                    if include_cim_fields:
-                        cef[k] = v
-
-            if use_event_id_sdi and "event_id" in item:
-                sdi = item["event_id"]
-            else:
-                if use_event_id_sdi and "event_id" not in item:
-                    logger.warning("use_event_id_sdi enabled but event_id missing, using hash")
-                input_str = UnicodeDammit(json.dumps(item)).unicode_markup.encode("utf-8")
-                if _get_fips_enabled():
-                    sdi = hashlib.sha256(input_str).hexdigest()
-                else:
-                    sdi = hashlib.md5(input_str).hexdigest()  # noqa: S324
-
-            severity = _get_splunk_severity(item)
-            spl_event_start = _get_event_start(item.get("_time"))
-            container_name = _get_splunk_title(item, container_name_prefix, container_name_values)
-
-            yield Container(
-                name=container_name,
-                severity=severity,
-                source_data_identifier=sdi,
-            )
-
-            if asset.remove_empty_cef:
-                cef = {k: v for k, v in cef.items() if v is not None}
-
-            yield Artifact(
-                cef=cef,
-                name="Field Values",
-                source_data_identifier=sdi,
-                severity=severity,
-                start_time=spl_event_start,
-            )
-
-            if count == asset.container_update_state and not is_poll_now:
-                state["start_time"] = item.get("_indextime")
-                count = 0
-            count += 1
-
-        except Exception as e:
-            logger.error("Error processing event: %s", e)
-            continue
-
-    if data and not is_poll_now:
-        state["start_time"] = data[-1].get("_indextime")
 
 
 if __name__ == "__main__":
